Fixed another issue with saving imported textures
Now compresses textures by default using DXT5. Compression is required to avoid the game hitting some internal limit it has on peg file size. The texture editor is almost ready for use. Need to add a few quality of life features and test it a bit more to make sure it's ready for use and then I'll make the first official release.